export utilities and fix multiple dates bugs
